[CAP-005] CommonUI, PresentationUtil모듈 생성및 의존 구조 개편
CommonUI : DSKit에 사용되는 기초 UI에서 고도화된 UI들이 위치합니다.
PresentationUtil : BaseVC, BaseVM, Coordinator와 같이 화면의 베이스가되는 기초 요소가 위치합니다.
CommonFeature : 익스텐션앱과 메인앱이 공유하는 Feature가 위치합니다.
