+++ Modules/Presentation/CommonUI/Source/Cell/SummaryKeywordCell.swift
@@ -1,6 +1,6 @@
 //
 //  SummaryKeywordCell.swift
-//  DSKit
+//  CommonUI
 //
 //  Created by choijunios on 8/22/24.
 //
@@ -9,6 +9,7 @@ import UIKit
 import Entity
 import RxSwift
 import RxCocoa
+import DSKit
 
 // MARK: Cell
 public class SummaryKeywordCell: UICollectionViewCell {
