+++ Modules/Presentation/MainAppFeatures/Source/MainScreen/Summaries/SubTabs/AllSummaryListView.swift
@@ -11,6 +11,7 @@ import RxCocoa
 import RxSwift
 import Entity
 import DSKit
+import CommonUI
 
 public class AllSummaryListView: UIView {
     typealias Cell = SummaryCell
