+++ Modules/Presentation/MainAppFeatures/Source/OnBoarding/SelectCategoryScreen/SelectMainCategoryVC.swift
@@ -11,6 +11,7 @@ import RxSwift
 import Entity
 import DSKit
 import BaseFeature
+import CommonUI
 
 public class SelectMainCategoryVC: BaseVC {
