+++ Modules/Presentation/MainAppFeatures/Source/OnBoarding/StartScreen/ClickToStartVC.swift
@@ -10,6 +10,7 @@ import RxCocoa
 import RxSwift
 import Entity
 import DSKit
+import CommonUI
 
 public class ClickToStartVC: UIViewController {
