+++ Modules/Presentation/PresentationUtil/Source/Extension/Swift+Extension/ResultExtensionForStream.swift
